% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM_trading_API_funs.R
\name{dm_create_advanced_target}
\alias{dm_create_advanced_target}
\title{Create Advanced Target}
\usage{
dm_create_advanced_target(
  gameId,
  title,
  price,
  items_to_buy,
  phase = "",
  paintSeed = 0,
  floatPartValue = "",
  verbose1 = FALSE
)
}
\arguments{
\item{gameId}{Character. The game ID of the item to create a target for.}

\item{title}{Character. The title of the item to create a target for.}

\item{price}{Numeric. The price of the item to create a target for.}

\item{items_to_buy}{Numeric. The number of items to buy.}

\item{phase}{Character. The phase of the item to create a target for.}

\item{paintSeed}{Numeric. The paint seed of the item to create a target for.}

\item{floatPartValue}{Character. The float part value of the item to create a target for.}

\item{verbose1}{Logical. Whether to print the response.}
}
\value{
A tibble containing the response.
}
\description{
This function creates an advanced metadata target on DMarket.
}
