% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM_trading_API_funs.R
\name{dm_create_offer}
\alias{dm_create_offer}
\title{Create an offer.}
\usage{
dm_create_offer(asset_id = NULL, price = NULL)
}
\arguments{
\item{asset_id}{The asset id.}

\item{price}{The price.}
}
\value{
A response.
}
\description{
Create an offer.
}
