% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM_trading_API_funs.R
\name{dm_create_target}
\alias{dm_create_target}
\title{Create target}
\usage{
dm_create_target(gameId, title, price, items_to_buy)
}
\arguments{
\item{gameId}{id of the game, CS:GO - a8db, Team Fortress 2 - tf2, Dota 2 - 9a92.}

\item{title}{title of the item AKA item name}

\item{price}{target price of the item}

\item{items_to_buy}{number of items to buy}
}
\value{
success of the operation
}
\description{
Creates target for buying items
}
