% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM_trading_API_funs.R
\name{dm_delete_target}
\alias{dm_delete_target}
\title{Delete Target on DMarket}
\usage{
dm_delete_target(target_id)
}
\arguments{
\item{target_id}{Character. The ID of the target to be deleted.}
}
\value{
A list with the response from the API.
}
\description{
This function deletes a target on DMarket based on the provided target ID.
}
