% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM_trading_API_funs.R
\name{dm_edit_offer}
\alias{dm_edit_offer}
\title{Edit offer.}
\usage{
dm_edit_offer(offer_id = NULL, asset_id = NULL, price = NULL)
}
\arguments{
\item{offer_id}{The offer ID.}

\item{asset_id}{The asset ID.}

\item{price}{The price.}
}
\value{
A response.
}
\description{
Edit offer.
}
