% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM_trading_API_funs.R
\name{dm_get_agg_price_all}
\alias{dm_get_agg_price_all}
\title{Get Aggregated Price for All Items}
\usage{
dm_get_agg_price_all(limit = 1000)
}
\arguments{
\item{limit}{Numeric. The maximum number of items to retrieve in a single API call.}
}
\value{
A tibble containing aggregated price information for all items.
}
\description{
This function retrieves aggregated price information for all items on DMarket.
If the number of items is greater than 1000, the function will make multiple API calls to retrieve the data.
}
