% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM_trading_API_funs.R
\name{dm_get_agg_price_wrap}
\alias{dm_get_agg_price_wrap}
\title{Get Aggregated Price for Items (Wrapper)}
\usage{
dm_get_agg_price_wrap(item_names)
}
\arguments{
\item{item_names}{Data frame. A data frame containing the names of the items to retrieve prices for.}
}
\value{
A tibble containing aggregated price information for the items.
}
\description{
This function retrieves aggregated price information for specified items on DMarket.
If the number of items is greater than 100, the function will make multiple API calls to retrieve the data.
}
