% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM_trading_API_funs.R
\name{dm_get_aggregated_price}
\alias{dm_get_aggregated_price}
\title{Get Aggregated Price for Items}
\usage{
dm_get_aggregated_price(item_name = NULL, Limit = "100", Offset = "0")
}
\arguments{
\item{item_name}{Character. The name of the item(s) to retrieve prices for.}

\item{Limit}{Character. Limit for number of returned items. Default is '100'.}

\item{Offset}{Character. Offset for response items. Default is '0'.}
}
\value{
A tibble containing aggregated price information for the items.
}
\description{
This function retrieves aggregated price information for specified items on DMarket.
}
