% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM_trading_API_funs.R
\name{dm_get_closed_targets}
\alias{dm_get_closed_targets}
\title{Get Closed Targets}
\usage{
dm_get_closed_targets(
  limit = NULL,
  offset = NULL,
  OrderDir = "desc",
  Presentation = "InventoryPresentationSimple"
)

dm_get_closed_targets(
  limit = NULL,
  offset = NULL,
  OrderDir = "desc",
  Presentation = "InventoryPresentationSimple"
)
}
\arguments{
\item{limit}{The limit.}

\item{offset}{The offset.}

\item{OrderDir}{The order direction.}

\item{Presentation}{The presentation.}
}
\value{
A tibble containing the response.

A response.
}
\description{
This function gets closed targets.

Get closed targets.
}
