% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM_trading_API_funs.R
\name{dm_get_customized_fees}
\alias{dm_get_customized_fees}
\title{Get Customized Fees}
\usage{
dm_get_customized_fees(
  GameID = NULL,
  offerType = "dmarket",
  limit = NULL,
  offset = NULL,
  add_extras = FALSE
)
}
\arguments{
\item{GameID}{Character. The game ID of the item to get.}

\item{offerType}{Character. The offer type of the item to get.}

\item{limit}{Integer. The number of items to get.}

\item{offset}{Integer. The offset of the items to get.}

\item{add_extras}{Logical. Whether to add extras.}
}
\value{
A tibble containing the response.
}
\description{
This function gets customized fees.
}
