% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM_trading_API_funs.R
\name{dm_get_my_balance}
\alias{dm_get_my_balance}
\title{Get my balance}
\usage{
dm_get_my_balance()
}
\value{
balance
}
\description{
Get user's balance
Balance in USD is of interest
}
\examples{
blnc <- get_my_balance()
}
