% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM_trading_API_funs.R
\name{dm_get_my_items}
\alias{dm_get_my_items}
\title{Get My Items}
\usage{
dm_get_my_items(
  GameID = NULL,
  title = NULL,
  limit = NULL,
  offset = NULL,
  orderBy = NULL,
  orderDir = NULL,
  treeFilters = NULL,
  currency = "USD",
  priceFrom = NULL,
  priceTo = NULL,
  classIds = NULL,
  cursor = NULL,
  add_extras = FALSE
)
}
\arguments{
\item{GameID}{Character. The game ID of the item to get.}

\item{title}{Character. The title of the item to get.}

\item{limit}{Integer. The number of items to get.}

\item{offset}{Integer. The offset of the items to get.}

\item{orderBy}{Character. The order of the items to get.}

\item{orderDir}{Character. The direction of the order.}

\item{treeFilters}{Character. The tree filters of the items to get.}

\item{currency}{Character. The currency of the items to get.}

\item{priceFrom}{Numeric. The price from of the items to get.}

\item{priceTo}{Numeric. The price to of the items to get.}

\item{classIds}{Character. The class IDs of the items to get.}

\item{cursor}{Character. The cursor of the items to get.}

\item{add_extras}{Logical. Whether to add extras.}
}
\value{
A tibble containing the response.
}
\description{
This function gets items from the user's inventory.
}
