% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM_trading_API_funs.R
\name{dm_get_offers_by_title}
\alias{dm_get_offers_by_title}
\title{Get Offers by Title}
\usage{
dm_get_offers_by_title(
  Title = NULL,
  Limit = 100,
  add_extras = FALSE,
  extra_col = NULL
)
}
\arguments{
\item{Title}{Character. The title of the item to retrieve offers for.}

\item{Limit}{Numeric. The maximum number of offers to retrieve.}

\item{add_extras}{Logical. Whether to add extra columns to the data.}

\item{extra_col}{Character. The extra columns to add to the data.}
}
\value{
A tibble containing offers by title.
}
\description{
This function retrieves offers by title on DMarket.
If item have extra columns, the function will add them to the data.
}
