% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM_trading_API_funs.R
\name{dm_get_targets}
\alias{dm_get_targets}
\title{Get Targets from DMarket}
\usage{
dm_get_targets(
  GameID = NULL,
  PriceFrom = NULL,
  PriceTo = NULL,
  Currency = NULL,
  Title = NULL,
  TargetID = NULL,
  Status = "TargetStatusActive",
  SortType = "UserTargetsSortTypeDefault",
  Offset = NULL,
  Limit = NULL,
  Cursor = NULL,
  is_active = NULL,
  add_extras = FALSE
)
}
\arguments{
\item{GameID}{Character. Game identifier on market.}

\item{PriceFrom}{Numeric. Minimum price for filtering.}

\item{PriceTo}{Numeric. Maximum price for filtering.}

\item{Currency}{Character. Currency code to use with price range filter.}

\item{Title}{Character. Filter user targets by asset's title.}

\item{TargetID}{Character. Filter by unique target identifier.}

\item{Status}{Character. Target status. Default is 'TargetStatusActive'.}

\item{SortType}{Character. Sort type for user targets. Default is 'UserTargetsSortTypeDefault'.}

\item{Offset}{Character. Offset for response items.}

\item{Limit}{Character. Limit for number of returned items.}

\item{Cursor}{Character. Cursor for the next page identifier.}

\item{is_active}{Logical. Alternative for BasicFilters.Status. Default is NULL.}

\item{add_extras}{Logical. Whether to add extra information. Default is FALSE.}
}
\value{
A tibble containing the list of targets.
}
\description{
This function retrieves a list of targets from DMarket based on various filters.
}
