% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM_trading_API_funs.R
\name{dm_get_targets_by_title}
\alias{dm_get_targets_by_title}
\title{Get targets by title.}
\usage{
dm_get_targets_by_title(game_id, title)
}
\arguments{
\item{game_id}{The game ID.}

\item{title}{The title.}
}
\value{
A response.
}
\description{
Get targets by title.
}
