% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM_trading_API_funs.R
\name{dm_get_targets_test}
\alias{dm_get_targets_test}
\title{Get targets from the sandbox.
 @param PriceTo Numeric. The price to of the item to get.}
\usage{
dm_get_targets_test(
  GameID = NULL,
  PriceFrom = NULL,
  PriceTo = NULL,
  Currency = NULL,
  Title = NULL,
  TargetID = NULL,
  Status = "TargetStatusActive",
  SortType = "UserTargetsSortTypeDefault",
  Offset = NULL,
  Limit = NULL,
  Cursor = NULL,
  is_active = NULL,
  attributes = FALSE
)
}
\description{
Get targets from the sandbox.
 @param PriceTo Numeric. The price to of the item to get.
}
