% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM_trading_API_funs.R
\name{dm_get_user_inventory}
\alias{dm_get_user_inventory}
\title{Get user inventory.}
\usage{
dm_get_user_inventory(
  limit = NULL,
  offset = NULL,
  BasicFilters.InMarket = TRUE
)
}
\arguments{
\item{limit}{The limit.}

\item{offset}{The offset.}

\item{BasicFilters.InMarket}{The basic filters.}
}
\value{
A response.
}
\description{
Get user inventory.
}
