% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM_trading_API_funs.R
\name{dm_get_user_offers}
\alias{dm_get_user_offers}
\title{Get user offers.}
\usage{
dm_get_user_offers(
  limit = NULL,
  offset = NULL,
  Status = "OfferStatusActive",
  GameID = NULL
)
}
\arguments{
\item{limit}{The limit.}

\item{offset}{The offset.}

\item{Status}{The status.}

\item{GameID}{The game ID.}
}
\value{
A response.
}
\description{
Get user offers.
}
