% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM_trading_API_funs.R
\name{dm_wrap_api_get}
\alias{dm_wrap_api_get}
\title{Wrap API Call}
\usage{
dm_wrap_api_get(api_fun, ...)
}
\arguments{
\item{api_fun}{Function. The API function to call.}

\item{...}{Additional arguments to pass to the API function.}
}
\value{
The data returned by the API function or an error message.
}
\description{
This helper function wraps an API call and returns the data or an error message.
}
