% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM_trading_API_funs.R
\name{mk_header}
\alias{mk_header}
\title{Create a header.}
\usage{
mk_header(
  method1,
  api_url_path,
  query_params = NULL,
  body = NULL,
  secret_key,
  nonce
)
}
\arguments{
\item{method1}{The method.}

\item{api_url_path}{The api url path.}

\item{query_params}{The query params.}

\item{body}{The body.}

\item{secret_key}{The secret key.}

\item{nonce}{The nonce.}
}
\value{
A header.
}
\description{
Create a header.
}
