% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM_trading_API_funs.R
\name{price_cols}
\alias{price_cols}
\title{price_cols}
\usage{
price_cols(data, nested_name)
}
\arguments{
\item{data}{A list. The data to unnest.}

\item{nested_name}{Character. The name of the nested list.}
}
\value{
unnseted data
}
\description{
This function unnests the price columns from the nested list.
}
