% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM_trading_API_funs.R
\name{q_parameter}
\alias{q_parameter}
\title{query parameter function}
\usage{
q_parameter(x = NULL)
}
\arguments{
\item{x}{parameter to use in API call}
}
\value{
query parameter
}
\description{
Creates query parameter for API call
uses quote function to get parameter name
and crete charecter string '{x_name}={x_value}'
}
